+++ core/src/main/java/io/fabric8/maven/core/util/kubernetes/KubernetesResourceUtil.java
@@ -108,12 +108,14 @@ public class KubernetesResourceUtil {
     public static final String API_APPS_VERSION = "apps/v1";
     public static final String JOB_VERSION = "batch/v1";
     public static final String OPENSHIFT_V1_VERSION = "apps.openshift.io/v1";
+    public static final String CRONJOB_VERSION = "batch/v1beta1";
     public static final ResourceVersioning DEFAULT_RESOURCE_VERSIONING = new ResourceVersioning()
             .withCoreVersion(API_VERSION)
             .withExtensionsVersion(API_EXTENSIONS_VERSION)
             .withAppsVersion(API_APPS_VERSION)
             .withOpenshiftV1Version(OPENSHIFT_V1_VERSION)
-            .withJobVersion(JOB_VERSION);
+            .withJobVersion(JOB_VERSION)
+            .withCronJobVersion(CRONJOB_VERSION);
 
     public static final HashSet<Class<?>> SIMPLE_FIELD_TYPES = new HashSet<>();
 
@@ -283,6 +285,8 @@ public class KubernetesResourceUtil {
             apiVersion = apiVersions.getJobVersion();
         } else if (Objects.equals(kind, "DeploymentConfig") && platformMode == PlatformMode.openshift) {
             apiVersion = apiVersions.getOpenshiftV1version();
+        } else if (Objects.equals(kind, "CronJob")){
+            apiVersion = apiVersions.getCronJobVersion();
         }
         addIfNotExistent(fragment, "apiVersion", apiVersion);