+++ CHANGELOG.md
@@ -18,6 +18,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix #1628: Adding back support for templates in kubernetes mode to HelmMojo; Fix for NPE in UndeployMojo
 * Fix 1591: Add support for custom resources creation via resource fragments
 * Fix 1636: Added support to choose configMap name
+* Fix 1630: Fix CronJob Version
 
 ### 4.1.0 (16-02-2019)
 * Fix 1578: Fmp is generating ImageChange triggers in the DeploymentConfig