+++ CHANGELOG.md
@@ -21,6 +21,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix #1636: Added support to choose configMap name
 * Fix #1630: Fix CronJob Version
 * Fix #1591: Add support for custom resources creation via resource fragments
+* Fix #1648: Job creation fails during fabric8:apply with error: forbidden
 
 ### 4.1.0 (16-02-2019)
 * Fix 1578: Fmp is generating ImageChange triggers in the DeploymentConfig