+++ CHANGELOG.md
@@ -28,6 +28,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix #1622: Additions to minimal profile
 * Feature #1661: Added Resource Versioning Support for RBAC
 * Fix 73: Added Service Port normalization
+* Bumped version to 4.2-SNAPSHOT
 
 ### 4.1.0 (16-02-2019)
 * Fix 1578: Fmp is generating ImageChange triggers in the DeploymentConfig